Separate cache and response from main
Generate a reqwest client from the cli

Remove redundant comments

Improve event handling structure

Break large functions into smaller ones

Prefer default_value_t instead of default_value

Remove redundant short and long attributes when inferred in clap

Replace #[clap] with #[arg]

Improve validation by returning an error instead of exiting the app
